backend/rest_api/requirements.txt
django==1.8.5
djangorestframework==3.2.4
mysqlclient==1.3.6
pillow==3.0.0
django-rest-swagger==0.3.4
